Moved homepage file contents to index.js to make it new landing page
