+++ components/NavBar.js
@@ -26,9 +26,6 @@ export default function NavBar() {
             <Link passHref href="/">
               <Nav.Link>Home</Nav.Link>
             </Link>
-            <Link passHref href="/homepage">
-              <Nav.Link>Sub Posts - Homepage</Nav.Link>
-            </Link>
             <Link passHref href="/posts/myPosts">
               <Nav.Link>My Posts</Nav.Link>
             </Link>
